Add class for /proc/net/sockstat, also support routines for it's record format
The /proc/net/sockstat file uses a record format that none of the other
routines currently in the library use.  So adding this code, along with
a new method in the SeqFileIO class and minor tweakd to the read_line
method, will make it easy to add other files in /proc/net/* that have
similar lines.