Add another demo script "show-active-timers"
The script "show-active-timers" runs through the info
in the "/proc/timer_list" file and extracts the pending
timers on the system.  Each one is printed with summary
info, indluding the full commandline of the process that
owners the timer.  So the script shows both how to pull
info from one "/proc" dataset and also how to do a
secondary pull using info from that file (the PID in
the timer file is used to get the command line).