proc2csv and proc2json pattern options and /proc/self/stat bug fix
Added second option to proc2csv and proc2json so you can explicitly
identify the /proc file architype file that should be used to parse
the target data file.

There was an "off by one" bug in the ProcSelfSTAT handler. The
argument number in the parse rules was skewed by one since the 21st
position was skipped by accident.