Modified version of /proc/self/cmdline handler
The /proc/self/cmdline handler now returns a parsed list of command
line arguments in addition to a reconstructed string with all the
arguments joined.  The split out list is required to accurately
reconstruct the original file since a given argument can contain
blanks.

Also, this commit includes updated verification files that reflect
the changes in the handler.