Changes to code libs, self/numa_maps handler and verify data output
In order to code a recreator for /proc/###/numa_maps files some
additional metadata about the original text files had to be captures
and recorded in the parsed version. The list of options name/value
pairson the line and the order in which the "nodes" are listed are
now available.  The changes required were to some core routines that
multiple handlers extend as well as the self/numa_maps specific parser.

And since there are new fields in the parsed data, the verification
output files (CSV and JSON) has to be regenerated as well.