Changes to /proc/interrupts handler and new verification test files
The descriptive field at the end of the /proc/interrupt lines is
preserved with all the spaces now.  Also, the width of the first
column, which is based on the number of interrupts on the system is
also returned in the set or parsed data.

And since the parsed data changed, updated version of the CSV/JSON
parsed output files (used by the verification test scripts) is also
included.