In /proc/timer_list handler, preserve the text version of the timer address
The address field in the timer subrecords in /proc/timer_list can be
text (and not a bracketed hex address).  This change stores the text
version of the field along with the parsed "long".  The change was needed
in order to avoid loosing non-numeric data that can be presented in the
file.