Functional, but not perfect, recreator for /proc/timer_list files
The kernel source that generates /proc/timer_list has a number of
"#ifdef" sections that conditionally include subrecords and even
some fields in records.  This version of the recreator isn't smart
enough to generate all the variants of the /proc/timer_list file
that may be possible on systems with different OS configurations.
But it works for systems that have all the features enabled.