Improvements to recreator tool for /proc/timer_list
There are a varierty of file formats possible for /proc/timer_list
depending on kernel options.  This update adds logic to figure out
some of those variants, but not all.  So it's an improvement but
there's more work to do before the tool can recreate /proc/timer_list
files on any given Linux system.