Fix parsing of trailing text fields in /proc/timer_stats files
The process name field in /proc/timer_stats records is a fixed length
field, sometime containing blanks.  The two text fields after that a
blank and parenthesis delimited.  So the algorithm to parse the fields
had to be customized since the stock parse rules aren't capable of
teasing the fields apart.