Transparency and 'order of records' changes to /proc/schedstat handler
A couple of changes were made to the ProcRootSCHEDSTAT handler with
this commit.

- The parsed records in the data file were returned in random order.
With this update, the records are returned in the order they appear
in the original data file.

- There's an additional meta-data field in the records that lists
the domain sub-records included.  The order those domains appeared
in the original file is also provided.

- The JSON and CSV verification files generated by "proc2csv" and
"proc2json" were also updated, since the changes to the handler
changed the order of records and added a new field.