Add general purpose recreator for all "ListOfTerms" based handlers
There are a number of "/proc" files that are just a list of terms,
one per line.  Those are all parsed with the "ListOfTerms" class
and can be recreated using the same code.  This commit adds a
general purpose method for recreating any of those files from its
parsed data.  And there also shell functions to create recreators
for each of the files that use handler.  The files supported are:
    /proc/net/ip_tables_targets
    /proc/net/ip_tables_matches
    /proc/net/ip_tables_names
    /proc/net/ip6_tables_targets
    /proc/net/ip6_tables_matches
    /proc/net/ip6_tables_names