Expose input data seq meta-data in read_twoline_logical_record()
In order to recreate the original input files from parsed data, the
handler for files like /proc/net/netstat and /proc/net/snmp need
to be able to figure out what order the fields appear in original
data file.  This change preserves that information in a dictionary
which can be scanned.