Bug fix, missing docs and verification udpates for /proc/crypto
The data presented in the /proc/crypto consistents of sets of data,
each of which has fields that determined the algorithm used by the
driver being described.  So the fields vary from group to group.
The ProcRootCRYPTO handler had parse rules in place to recognize
all but one possible field.  This update adds a parse rule for
"seedsize" lines correct that oversight.

Also, since that change effects the output generated the JSON and
CSV verification files for that driver had to be updated.

Finally, the README.ProcRootHandlers doc file didn't include the
kernel snippets that show how the lines in /proc/crypto were
generated.  An expanded version of that documentation is also part
of this commit.