Correct "fixpos_hits" tracking in FixedWhitespaceDelimRecs class
Previously, if all the selection criteria matched for a parse rule
the "fixpos_hits" dictionary would get an entry saying that rule
was used even if the number of parameters on the line was too few
to actually get raw data to parse.  This change moves the logic
to add entries to the dictionary inside the block bracketed by a
check for the number of entries on the line.