Enhancements to /proc/version handler and new verification files
The ProcRootVERSION handler wasn't pulling out all the fields in
the version string.  If was only getting the fields listed as
variables in the seq_printf() call in the kernel.  But the format
string itself is dynamic on a per-system basis.  Even though those
fields only change per release, it's worth extracting them.

Also, since new fields are parsed out of the input file, new JSON
and CSV output files used by the "verify*" tools had to be
generated.  Those are included in this commit as well.