Add special case cose in recreate-proc-file for /proc/self/io
Copying the file /proc/self/io is problematic...  When doing a
normal copy that increments I/O counts, the resulting copy often
has stray garbage at the end.  That's presumably since the
process copying the I/O counter file in /proc is the process whose
data is contained in the file.  Ugh...

So the "recreate-proc-file" script checks for that file and does
a two-step copy.