One fix for /proc/self/sched, one for /proc/self/smaps
The /proc/self/sched handler fix is a partial one.  It changes the
handler so that it can parse files generated by newer kernels, which
have 4 additional fields.  But additional changes will be required
to capture and pass along meta-data about the formatting of the input
file since the width of the first text field changed at some point.
The recreator for the data file can't determine how to generate exact
matches for the original data without knowing how many characters to
use for the text field.

The /proc/self/smaps handler fix is trivial.  The list of two letter
"VmFlags" have a trailing space.  The code needs to preserve that.