Handler, recreator and verification file updates for /proc/self/sched
The handler passes along a decorative, but useful, field in the
parsed data.  That allows the recreator to conditionally format the
width of the descritor field (the first column) to the proper width.

And since the parsed output changes, the JSON/CSV output files used
by the verification scripts needed to be udpated as well.