File /proc/self/stat has additional fields in later kernels
Eight new fields may be in the /proc/self/stat file depending on the
kernel patch level.  The ProcSelfSTAT handler is able to parse either
version of the file now.  The new fields are also defined in the
ProcFieldConstants.py file.

And the alternate version of that part of the kernel source code is
now in the README.ProcSelfHandlers doc file.