Changes for new "pylint"
The config file changed to disable warnings about bad "before" and
"after" spaces, since the fixes would make the code too ugly.  Also
change the format of the output lines to include error numbers and
to make the look more like the older version.

And some minor tweaks to remove trailing spaces in "proc2json" which
it's not tracking (which is a nice improvement).