Ubuntu 14 fixes /proc/self/status handler and some pylint updates
In Ubuntu 14 there's a new field in /proc/self/status which the
handler wasn't looking for.  This update adds a new field, so the
JSON/CSV verification test data files are also updated.

The Ubuntu 14 "pylint" also checks for a few new things so the
ProcSelfHandlers.py and ProcFieldConstants.py file, which had to
be changed for the functional update, were cleaned up as well.