Multiple changes bundled together (unfortunately)
There are changes here to disambiguate the strings associated with
the symbol fields names defined in ProcFieldConstants.py.  And that
led to changes in code that referenced a couple of field constants
that were duplicates (which were collapsed together).  And finally,
the JSON/CSV parsed output files used by verification scripts had
to be updated for some handlers as well.

In addition, there are changes to the handler and the recreator for
the file /proc/self/sched.  Those changes were required to deal with
Ubuntu 14 versions of the file.  They have some extra NUMA related
lines that were in not in Ubuntu 12 when the code was originally
developed.