Handle variable formats (kernel dependent) for /proc/net/netlink
The "dump" field in /proc/net/netlink can be adjustable length
base 10 decimal number, or a fixed 16 character hex string.  The
changes in this commit make the handler and the recreator for that
file able to deal with either.  Also, the JSON/CSV files used by
verification scripts were updated to reflect the new field added
to provide the meta-data needed to recreate the original file from
the parsed data.