Ubuntu 14 tweak to /proc/timer_stats file code
Ubuntu 14 includes an extra line in /proc/timer_stats which was
ignored by the handler.  This update adds a new field to pick up
the "collections" setting, includes new JSON/CSV verification
files, and adds code to the recreator to conditionally generate the
new line if was presented in the source data file.