Change parse method for /proc/net/ptype
The data in /proc/net/ptype is in fixed columns, but the parsing
code was using the FixedWhitespaceDelimRecs base class with custom
logic to figure out how to back into properly parsed fields.  This
update switchs the handler to use the FixedColumnRecs base class
to fix a bug (and simplify the code).

Also since the parsed data now can have trailing whitespace, the
JSON and CSV files used by the verifcation scripts was updated.

And finally, the recreator for /proc/net/ptype needed a minor change
as well.