Fedora 21 changes to /proc/zoneinfo file centric code/data
Fedora 21 includes fields that aren't in the primary development OS
for this code.  The udpates included in this commit add support for
those extra fields in the ProcRootZONEINFO handler.

Also, the JSON/CSV data files are updated to include the new fields
so that the verification scripts will work.

Finally, the new fields (and the associated output formats) were
added to the recereator for /proc/zoneinfo as well.