Move recreator checks into "regentest-all-proc-files" driver
The "regentest-all-proc-files" script relied on the script it called
for each input file to check whether or not there was a recreator
module available, whether the requested /proc file existed on the
OS where it's being run, and to determine if the user running the
script had read priv's to the file.

That made the output messier and full of "error" messages that were
normal/expected.  So the checking was moved to the driver script to
control the messaages better.