Debian 7 fix for /proc/timer_list recreator
The number of blank lines between each group of lines that make up a
logical record is very difficult to describe logically.  So the
recreator for /proc/timer_list now just relies on meta-data from the
parse to figure out when extra blanks are needed to recreate the
original file.