Add support for bluetooth centric data files in /proc/net
The files /proc/net/bnep, /proc/net/hci, /proc/net/l2cap, and
/proc/net/sco all use the same file format.  There's one generic
bluetooth data format handler that they all share.  The only
reason for the separate handlers at all is to create a mapping
from the filenames to the code to parse the file.

The sample output files (meaning the JSON and CSV's) for all of
them are include so that the verification script will work.

And there's a general purpose bluetooth data format recreator
function too.  There are separate shell recreators that share the
common code, again in order to setup mapping from filesnames to
code so that lookup functions will work.