Bug fix for /proc/mdstat, only for "inactive" ldev use cases
When the status of a logical device is "inactive", then two fields
which are normally present are blank.  This update changes the
handler that parses the data to handle skipping the missing field.
And there's another change to the "regenerator" to make it smart
enough to skip writing two status fields to the file it creates.