Smarter "end of group" code for /proc/self/smaps handler
A sample /proc/###/smaps file from a Fedora system I received was
not parsing.  It turns out the "last field" coded found in the kernel
code use to model the parser hasn't always been used.  The new code
uses more flexible logic for deciding when a new memory segment
starts in the file.